Add files via upload
This is an example of a login screen that is utilizing a SQL database. The code will not run if the JConnector is not setup, but definitely look at the coding to see how the SQL queries are utilized. The login screen in this example will only match the password entered to the password of the correct username that is entered.