Add files via upload
These files are the billing window that I completed. The classes currently work by itself. The main menu of the RIS should feature a billing button that calls billing.main(physician); the String physician value should be the name gathered from the login credentials. As of now, it is pseudo-medical doctor Dr. Bryce Pain which is a static value.